[#142983763] - add save logic for reminder lists

+++ KeithLink/Entree.Migrations/Scripts/AppData/List/Tables/CustomListHeaders/CustomListHeaders_0031_2017.04.00_up.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [List].[CustomListHeaders]
 (
-	[Id]				INT					NOT NULL	PRIMARY KEY IDENTITY(1,1),
+	[Id]				BIGINT				NOT NULL	PRIMARY KEY IDENTITY(1,1),
 	[UserId]			UNIQUEIDENTIFIER	NULL,
 	[BranchId]			CHAR(3)				NOT NULL,
 	[CustomerNumber]	CHAR(6)				NOT NULL,
