+++ KeithLink/Entree.Migrations/Scripts/AppData/List/Tables/InventoryValuationListHeaders/InventoryValuationListHeaders_0030_2017.04.00_up.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [List].[InventoryValuationListHeaders]
 (
-	[Id]                INT             NOT NULL    PRIMARY KEY IDENTITY(1,1),
+	[Id]                BIGINT          NOT NULL    PRIMARY KEY IDENTITY(1,1),
 	[BranchId]          CHAR(3)         NOT NULL,
 	[CustomerNumber]    CHAR(6)         NOT NULL,
 	[Name]              NVARCHAR(150)   NOT NULL,
